+++ R/zzz.R
@@ -34,6 +34,7 @@ utils::globalVariables(c(
   "cpp_phylo_knn_parallel", "cpp_phylo_knn_single",
   "cpp_radius_parallel", "cpp_radius_single",
   "cpp_random_parallel", "cpp_random_single", "cpp_order_parallel",
+  "cpp_knn_order", "cpp_kncn_order", "calc_rao", "calc_rao_traits",
   "cpp_wavefront_parallel", "cpp_wavefront_single",
   "interpolate_at_coverage",
   "cpp_knn_hill_coverage_parallel",